Refactor to put OutputEstimate in its own class.
Had to move the read_run_mask to the TEMUtilityFunctions.cpp
