+++ src/TEM.cpp
@@ -67,6 +67,7 @@
 
 #include "inc/timeconst.h"
 #include "../include/ArgHandler.h"
+#include "../include/OutputEstimate.h"
 #include "TEMLogger.h"
 #include "TEMUtilityFunctions.h"
 #include "../include/Runner.h"
@@ -94,9 +95,6 @@ void pp_2dvec(const std::vector<std::vector<int> > & vv);
 // draft - reading new-style co2 file
 std::vector<float> read_new_co2_file(const std::string &filename);
 
-// draft - reading in a 2D run mask...
-std::vector< std::vector<int> > read_run_mask(const std::string &filename);
-
 /** Enables a 'floating point exception' mask that will make the program crash
  *  when a floating point number is generated (and or operated upon).
  *
@@ -121,312 +119,6 @@ void enable_floating_point_exceptions() {
   #endif
 
 }
-
-struct NetcdfOutputTypes {
-  double daily;
-  double monthly;
-  double yearly;
-  NetcdfOutputTypes():daily(0), monthly(0), yearly(0) {}
-};
-
-struct JsonOutputTypes {
-  double archive;
-  double daily;
-  double monthly;
-  double yearly;
-
-  double jcoef_archive;
-  double jcoef_daily;
-  double jcoef_monthly;
-  double jcoef_yearly;
-
-  JsonOutputTypes(double _jca, double _jcd, double _jcm, double _jcy):
-    archive(0), daily(0), monthly(0), yearly(0),
-    jcoef_archive(_jca), jcoef_daily(_jcd), jcoef_monthly(_jcm), jcoef_yearly(_jcy) {}
-};
-
-/** Holds the ouput volume estimate for a run-stage. Notes:
- *   - sizes are in bytes
- *   - for all years represented by runyears
- *   - for one cell
- */
-struct StageOutputEstimate {
-  std::string name;
-  int runyears;
-  NetcdfOutputTypes nc_out;
-  JsonOutputTypes json_out;
-  StageOutputEstimate(
-      const std::string& _name, int _yrs,
-      int _jca, int _jcd, int _jcm, int _jcy):
-    name(_name),
-    runyears(_yrs),
-    nc_out(),
-    json_out(_jca, _jcd, _jcm, _jcy){}
-};
-
-class OutputEstimate {
-
-  /** Returns a 'human readable' size string with SI suffix */
-  std::string hsize(double size) {
-    int i = 0;
-    const std::string units[] = {"B","kB","MB","GB","TB","PB","EB","ZB","YB"};
-    while (size > 1024) {
-      size /= 1024;
-      i++;
-    }
-    std::stringstream ss;
-    ss.precision(0);
-    (size < 0.5) ? size = 0 : size=size;
-    ss << fixed << size << " " << units[i];
-    std::string size_string = ss.str();
-    return size_string;
-  }
-
-
-  int active_cells;
-  std::vector<StageOutputEstimate> stage_output_estimates;
-
-public:
-
-  OutputEstimate(const ModelData& md, bool calmode) {
-
-    stage_output_estimates = boost::assign::list_of
-        // Table of coefficients for bytes per year of run-time for
-        // each stage and different types of json output (archive, daily, etc).
-        (StageOutputEstimate("pr", md.pr_yrs, 22000,  41000,  200000,  20000))
-        (StageOutputEstimate("eq", md.eq_yrs, 56000,  41000,  240000,  24000))
-        (StageOutputEstimate("sp", md.sp_yrs, 15000,  36000,  199000,  16000))
-        (StageOutputEstimate("tr", md.tr_yrs, 15000,  36000,  199000,  16000))
-        (StageOutputEstimate("sc", md.sc_yrs, 15000,  36000,  199000,  16000));
-
-    // Open the run mask (spatial mask) and count all the non-zero cells.
-    std::vector< std::vector<int> > run_mask = read_run_mask(md.runmask_file);
-
-    this->active_cells = 0;
-
-    // Use a few type definitions to save some typing.
-    typedef std::vector<int> vec;
-    typedef std::vector<vec> vec2D;
-
-    vec2D::const_iterator row;
-    vec::const_iterator col;
-    for (row = run_mask.begin(); row != run_mask.end() ; ++row) {
-      for (col = row->begin(); col != row->end(); ++col) {
-        bool mask_value = *col;
-        if (mask_value) {this->active_cells++;}
-      }
-    }
-
-    std::cout << "Number of active cells: " << this->active_cells << std::endl;
-
-    std::vector<StageOutputEstimate>::iterator itr;
-    for (itr = stage_output_estimates.begin(); itr != stage_output_estimates.end(); itr++){
-
-      std::cout << "Estimating output for stage: " << itr->name << std::endl;
-      // handle all the calibration/json stuff
-      if (calmode) {
-        // cal mode means at least yearly and daily
-        std::cout << "  Estimating JSON CALIBRATION output" << std::endl;
-        itr->json_out.yearly = itr->runyears * itr->json_out.jcoef_yearly;
-        itr->json_out.daily = itr->runyears * itr->json_out.jcoef_daily;
-
-        // setting in config file
-        if (md.output_monthly) {
-          itr->json_out.monthly = itr->runyears * itr->json_out.jcoef_monthly;
-        } else {
-          itr->json_out.monthly = 0;
-        }
-        // could be an overestimate if user does not have monthly enabled too?
-        if (md.tar_caljson) {
-          itr->json_out.archive = itr->runyears * itr->json_out.jcoef_archive;
-        } else {
-          itr->json_out.archive = 0;
-        }
-      }
-
-      // Handle the NetCDF stuff...
-      double D_est = 0; double M_est = 0; double Y_est = 0;
-
-      // yearly
-      std::cout << "  Estimating YEARLY NC output for stage: " << itr->name << std::endl;
-      std::map<std::string, OutputSpec>::const_iterator map_itr;
-      for(map_itr = md.yearly_netcdf_outputs.begin(); map_itr != md.yearly_netcdf_outputs.end(); ++map_itr ){
-
-        double output_estimate = 8;
-        OutputSpec os = map_itr->second;
-
-        (os.pft) ? (output_estimate *= NUM_PFT) : output_estimate *= 1;
-        (os.compartment) ? (output_estimate *= NUM_PFT_PART) : output_estimate *= 1;
-        (os.layer) ? (output_estimate *= MAX_SOI_LAY) : output_estimate *= 1;
-        (os.yearly) ? (output_estimate *= (1 * itr->runyears)) : output_estimate *= 1;
-
-        Y_est += output_estimate;
-      }
-      map_itr = md.yearly_netcdf_outputs.end();
-
-      // monthly
-      std::cout << "  Estimating MONTHLY NC output for stage: " << itr->name << std::endl;
-      for(map_itr = md.monthly_netcdf_outputs.begin(); map_itr != md.monthly_netcdf_outputs.end(); ++map_itr ){
-
-        double output_estimate = 8;
-        OutputSpec os = map_itr->second;
-
-        (os.pft) ? (output_estimate *= NUM_PFT) : output_estimate *= 1;
-        (os.compartment) ? (output_estimate *= NUM_PFT_PART) : output_estimate *= 1;
-        (os.layer) ? (output_estimate *= MAX_SOI_LAY) : output_estimate *= 1;
-        (os.monthly) ? (output_estimate *= (12 * itr->runyears)) : output_estimate *= 1;
-
-        M_est += output_estimate;
-      }
-      map_itr = md.monthly_netcdf_outputs.end();
-
-      // daily
-      std::cout << "  Estimating DAILY NC output for stage: " << itr->name << std::endl;
-      for(map_itr = md.daily_netcdf_outputs.begin(); map_itr != md.daily_netcdf_outputs.end(); ++map_itr ){
-
-        double output_estimate = 8;
-        OutputSpec os = map_itr->second;
-
-        (os.pft) ? (output_estimate *= NUM_PFT) : output_estimate *= 1;
-        (os.compartment) ? (output_estimate *= NUM_PFT_PART) : output_estimate *= 1;
-        (os.layer) ? (output_estimate *= MAX_SOI_LAY) : output_estimate *= 1;
-        (os.daily) ? (output_estimate *= (365 * itr->runyears)) : output_estimate *= 1;
-
-        D_est += output_estimate;
-      }
-      map_itr = md.daily_netcdf_outputs.end();
-
-      itr->nc_out.yearly = Y_est;
-      itr->nc_out.monthly = M_est;
-      itr->nc_out.daily = D_est;
-
-      std::cout << std::endl;
-    }
-
-    //std::cout << "SET BREAKPOINT HERE\n";
-  }
-
-  void print_estimate(){
-
-    std::stringstream ss;
-    std::vector<StageOutputEstimate>::iterator itr;
-
-    ss << "-- calibration json output data volume estimates: " << hsize(json_total()) << std::endl;
-    ss << std::setw(10) << " "
-       << std::setw(10) << "run years"
-       << std::setw(10) << "archive"
-       << std::setw(10) << "daily"
-       << std::setw(10) << "monthly"
-       << std::setw(10) << "yearly"
-       << std::endl;
-    for (itr = stage_output_estimates.begin(); itr != stage_output_estimates.end(); itr++) {
-      ss << std::setw(10) << (*itr).name
-         << std::setw(10) << (*itr).runyears
-         << std::setw(10) << hsize((*itr).json_out.archive)
-         << std::setw(10) << hsize((*itr).json_out.daily)
-         << std::setw(10) << hsize((*itr).json_out.monthly)
-         << std::setw(10) << hsize((*itr).json_out.yearly)
-         << std::endl;
-    }
-    itr = stage_output_estimates.end();
-    std::cout << ss.str();
-    std::cout << std::endl;
-    ss.str("");
-    ss.clear(); // clear state flags
-
-    ss << "-- netcdf output data volume estimate: " << hsize(netcdf_total()) << std::endl;
-    ss << std::setw(10) << " "
-       << std::setw(10) << "run years"
-       << std::setw(10) << "daily"
-       << std::setw(10) << "monthly"
-       << std::setw(10) << "yearly"
-       << std::endl;
-
-    for (itr = stage_output_estimates.begin(); itr != stage_output_estimates.end(); itr++) {
-      ss << std::setw(10) << (*itr).name;
-      ss << std::setw(10) << (*itr).runyears;
-      ss << std::setw(10) << hsize((*itr).nc_out.daily);
-      ss << std::setw(10) << hsize((*itr).nc_out.monthly);
-      ss << std::setw(10) << hsize((*itr).nc_out.yearly);
-      ss << std::endl;
-    }
-    itr = stage_output_estimates.end();
-    std::cout << ss.str();
-    std::cout << std::endl;
-    ss.str("");
-    ss.clear(); // clear state flags
-
-    std::cout << "Cell Total: " << hsize(this->cell_total()) << std::endl;
-    std::cout << "All cells: " << hsize(this->all_cells_total()) << std::endl;
-  }
-
-  double netcdf_total() {
-    double t = 0;
-    std::vector<StageOutputEstimate>::iterator itr;
-    for (itr = stage_output_estimates.begin(); itr != stage_output_estimates.end(); itr++) {
-      t += (*itr).nc_out.daily;
-      t += (*itr).nc_out.monthly;
-      t += (*itr).nc_out.yearly;
-    }
-    return t;
-  }
-
-  double json_total() {
-    // This will probably over-estimate, as the yearly/monthly/daily outputs
-    // are deleted for each stage. Maybe should find the stage with the max
-    // runyears and use that for the yearly/monthly/daily estimates.
-    double t = 0;
-    std::vector<StageOutputEstimate>::iterator itr;
-    for (itr = stage_output_estimates.begin(); itr != stage_output_estimates.end(); itr++) {
-      t += (*itr).json_out.archive;
-      t += (*itr).json_out.daily;
-      t += (*itr).json_out.monthly;
-      t += (*itr).json_out.yearly;
-    }
-    return t;
-  }
-
-  double cell_total() {
-    double t = 0;
-      t += json_total();
-      t += netcdf_total();
-    return t;
-  }
-
-  double all_cells_total() {
-    return active_cells * (json_total() + netcdf_total());
-  }
-
-  /** (UNTESTED!) Given a string like "1.5 MB" should return a double (bytes). */
-  int hsize2bytes(const std::string& sizestr) {
-    // last 2 chars
-    std::string hrunit = sizestr.substr(sizestr.size()-2, std::string::npos); // last 2 chars
-
-    // first bit of string (till 2nd to last char), converted to dbl
-    double size = atof(sizestr.substr(0, sizestr.size()-2).c_str());
-
-    //assert( (size > 0 && size < ??) & "INVALID SIZE! CAN'T CONVERT!");
-
-    std::vector<std::string> units = boost::assign::list_of("B")("kB")("MB")("GB")("TB")("PB")("EB")("ZB")("YB");
-
-    //assert( !(std::find(units.begin(), units.end(), hrunit) != units.end()) & "INVALID UNITS!!" );
-
-    int i = 0;
-    while ( !(hrunit.compare(units[i])) == 0 )  {
-      size *= 1024;
-      i++;
-    }
-    return size; // in bytes
-  }
-
-// ideas for  API functions
-//  oe.per_grid_cell()
-//  oe.per_year()
-//  oe.print_table()
-
-};
-
-
 
 ArgHandler* args = new ArgHandler();
 
@@ -484,7 +176,7 @@ int main(int argc, char* argv[]){
   BOOST_LOG_SEV(glg, debug) << "NEW STYLE: Going to run space-major over a 2D area covered by run mask...";
 
   // Open the run mask (spatial mask)
-  std::vector< std::vector<int> > run_mask = read_run_mask(modeldata.runmask_file);
+  std::vector< std::vector<int> > run_mask = temutil::read_run_mask(modeldata.runmask_file);
 
   // Make some convenient handles for later...
   std::string eq_restart_fname = modeldata.output_dir + "restart-eq.nc";
@@ -884,67 +576,6 @@ void pp_2dvec(const std::vector<std::vector<int> > & vv) {
     std::cout << std::endl;
   }
 }
-
-/** rough draft for reading a run-mask (2D vector of ints)
-*/
-std::vector< std::vector<int> > read_run_mask(const std::string &filename) {
-  int ncid;
-  
-  BOOST_LOG_SEV(glg, debug) << "Opening dataset: " << filename;
-  temutil::nc( nc_open(filename.c_str(), NC_NOWRITE, &ncid) );
-  
-  BOOST_LOG_SEV(glg, debug) << "Find out how much data there is...";
-  int yD, xD;
-  size_t yD_len, xD_len;
-
-  temutil::nc( nc_inq_dimid(ncid, "Y", &yD) );
-  temutil::nc( nc_inq_dimlen(ncid, yD, &yD_len) );
-
-  temutil::nc( nc_inq_dimid(ncid, "X", &xD) );
-  temutil::nc( nc_inq_dimlen(ncid, xD, &xD_len) );
-
-  BOOST_LOG_SEV(glg, debug) << "Allocate a 2D run-mask vector (y,x). Size: (" << yD_len << ", " << xD_len << ")";
-  std::vector< std::vector<int> > run_mask(yD_len, std::vector<int>(xD_len));
-  
-  BOOST_LOG_SEV(glg, debug) << "Read the run flag data from the file into the 2D vector...";
-  int runV;
-  temutil::nc( nc_inq_varid(ncid, "run", &runV) );
-
-  BOOST_LOG_SEV(glg, debug) << "Grab one row at a time";
-  BOOST_LOG_SEV(glg, debug) << "(need contiguous memory, and vector<vector> are not contiguous)";
-
-  std::vector< std::vector<int> >::iterator row;
-  for (row = run_mask.begin();  row != run_mask.end(); ++row) {
-
-    int rowidx = row - run_mask.begin();
-
-    // specify start indices for each dimension (y, x)
-    size_t start[2];
-    start[0] = rowidx;    // Y
-    start[1] = 0;         // X
-
-    // specify counts for each dimension
-    size_t count[2];
-    count[0] = 1;         // one row
-    count[1] = xD_len;    // all data
-    
-    std::vector<int> rowdata(xD_len);
-
-    temutil::nc( nc_get_vara_int(ncid, runV, start, count, &rowdata[0] ) );
-  
-    run_mask[rowidx] = rowdata;
-    
-  }
-  
-  temutil::nc( nc_close(ncid) );
-
-  //pp_2dvec(run_mask);
-
-  BOOST_LOG_SEV(glg, debug) << "Return the vector...";
-  return run_mask;
-
-}
-
 
 /** rough draft for reading new-style co2 data
 */
